pyproject.toml
[tool.poetry]
name = "graphql first app"
version = "0.1.0"
description = ""
authors = ["Dakoh Wisdom <wiizzydreadmill@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
strawberry-graphql = "^0.205.0"
fastapi = "^0.103.1"
mongoengine = "^0.27.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.23.2"
certifi = "^2023.7.22"
bcrypt = "^4.0.1"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
